feat: move market information to context
